+++ src/YtdlCore.ts
@@ -1,7 +1,8 @@
-type YTDL_Constructor = YTDL_DownloadOptions & {
+type YTDL_Constructor = Omit<YTDL_DownloadOptions, 'format'> & {
     debug?: boolean;
 };
 
+import { fetch } from 'undici';
 import { PassThrough } from 'stream';
 import miniget from 'miniget';
 import m3u8stream, { parseTimestamp } from 'm3u8stream';
@@ -284,6 +285,7 @@ class YtdlCore {
     public clients: Array<YTDL_ClientTypes> | undefined = undefined;
     public disableDefaultClients: boolean = false;
     public oauth2: OAuth2 | undefined;
+    public notParsingHLSFormat: boolean = false;
 
     /* Format Selection Options */
     public quality: YTDL_ChooseFormatOptions['quality'] | undefined = undefined;
@@ -358,7 +360,7 @@ class YtdlCore {
     private initializeHtml5PlayerCache() {
         const HTML5_PLAYER = FileCache.get<{ playerUrl: string }>('html5Player');
 
-        if (!HTML5_PLAYER) {
+        if (!HTML5_PLAYER && !process.env._YTDL_DISABLE_HTML5_PLAYER_CACHE) {
             Logger.debug('To speed up processing, html5Player and signatureTimestamp are pre-fetched and cached.');
             getHtml5Player('dQw4w9WgXcQ', {}).then(async ({ playerUrl, path }) => {
                 if (!playerUrl) {
@@ -380,7 +382,11 @@ class YtdlCore {
         }
     }
 
-    constructor({ lang, requestOptions, rewriteRequest, agent, poToken, visitorData, includesPlayerAPIResponse, includesNextAPIResponse, includesOriginalFormatData, includesRelatedVideo, clients, disableDefaultClients, oauth2, quality, filter, excludingClients, includingClients, range, begin, liveBuffer, highWaterMark, IPv6Block, dlChunkSize, debug }: YTDL_Constructor = {}) {
+    constructor({ lang, requestOptions, rewriteRequest, agent, poToken, visitorData, includesPlayerAPIResponse, includesNextAPIResponse, includesOriginalFormatData, includesRelatedVideo, clients, disableDefaultClients, oauth2, notParsingHLSFormat, quality, filter, excludingClients, includingClients, range, begin, liveBuffer, highWaterMark, IPv6Block, dlChunkSize, debug, disableFileCache }: YTDL_Constructor = {}) {
+        /* Other Options */
+        process.env.YTDL_DEBUG = (debug ?? false).toString();
+        process.env._YTDL_DISABLE_FILE_CACHE = (disableFileCache ?? false).toString();
+
         /* Get Info Options */
         this.lang = lang || 'en';
         this.requestOptions = requestOptions || {};
@@ -392,6 +398,7 @@ class YtdlCore {
         this.includesRelatedVideo = includesRelatedVideo ?? true;
         this.clients = clients || undefined;
         this.disableDefaultClients = disableDefaultClients ?? false;
+        this.notParsingHLSFormat = notParsingHLSFormat ?? false;
         this.setPoToken(poToken);
         this.setVisitorData(visitorData);
         this.setOAuth2(oauth2);
@@ -410,41 +417,39 @@ class YtdlCore {
         this.IPv6Block = IPv6Block || undefined;
         this.dlChunkSize = dlChunkSize || undefined;
 
-        /* Debug Options */
-        process.env.YTDL_DEBUG = (debug ?? false).toString();
-
         this.automaticallyGeneratePoToken();
         this.initializeHtml5PlayerCache();
     }
 
     private setupOptions(options: YTDL_DownloadOptions) {
-        options.lang ??= this.lang;
-        options.requestOptions ??= this.requestOptions;
-        options.rewriteRequest ??= this.rewriteRequest;
-        options.agent ??= this.agent;
-        options.poToken ??= this.poToken;
-        options.visitorData ??= this.visitorData;
-        options.includesPlayerAPIResponse ??= this.includesPlayerAPIResponse;
-        options.includesNextAPIResponse ??= this.includesNextAPIResponse;
-        options.includesOriginalFormatData ??= this.includesOriginalFormatData;
-        options.includesRelatedVideo ??= this.includesRelatedVideo;
-        options.clients ??= this.clients;
-        options.disableDefaultClients ??= this.disableDefaultClients;
-        options.oauth2 ??= this.oauth2;
+        options.lang = options.lang || this.lang;
+        options.requestOptions = options.requestOptions || this.requestOptions;
+        options.rewriteRequest = options.rewriteRequest || this.rewriteRequest;
+        options.agent = options.agent || this.agent;
+        options.poToken = options.poToken || this.poToken;
+        options.visitorData = options.visitorData || this.visitorData;
+        options.includesPlayerAPIResponse = options.includesPlayerAPIResponse || this.includesPlayerAPIResponse;
+        options.includesNextAPIResponse = options.includesNextAPIResponse || this.includesNextAPIResponse;
+        options.includesOriginalFormatData = options.includesOriginalFormatData || this.includesOriginalFormatData;
+        options.includesRelatedVideo = options.includesRelatedVideo || this.includesRelatedVideo;
+        options.clients = options.clients || this.clients;
+        options.disableDefaultClients = options.disableDefaultClients || this.disableDefaultClients;
+        options.oauth2 = options.oauth2 || this.oauth2;
+        options.notParsingHLSFormat = options.notParsingHLSFormat || this.notParsingHLSFormat;
 
         /* Format Selection Options */
-        options.quality ??= this.quality || undefined;
-        options.filter ??= this.filter || undefined;
-        options.excludingClients ??= this.excludingClients || [];
-        options.includingClients ??= this.includingClients || 'all';
+        options.quality = options.quality || this.quality || undefined;
+        options.filter = options.filter || this.filter || undefined;
+        options.excludingClients = options.excludingClients || this.excludingClients || [];
+        options.includingClients = options.includingClients || this.includingClients || 'all';
 
         /* Download Options */
-        options.range ??= this.range || undefined;
-        options.begin ??= this.begin || undefined;
-        options.liveBuffer ??= this.liveBuffer || undefined;
-        options.highWaterMark ??= this.highWaterMark || undefined;
-        options.IPv6Block ??= this.IPv6Block || undefined;
-        options.dlChunkSize ??= this.dlChunkSize || undefined;
+        options.range = options.range || this.range || undefined;
+        options.begin = options.begin || this.begin || undefined;
+        options.liveBuffer = options.liveBuffer || this.liveBuffer || undefined;
+        options.highWaterMark = options.highWaterMark || this.highWaterMark || undefined;
+        options.IPv6Block = options.IPv6Block || this.IPv6Block || undefined;
+        options.dlChunkSize = options.dlChunkSize || this.dlChunkSize || undefined;
 
         if (!this.oauth2 && options.oauth2) {
             Logger.warning('The OAuth2 token should be specified when instantiating the YtdlCore class, not as a function argument.');